[refactor] refactoring Competition View & ViewModel
