+++ Git-Challenges/Let's Git it!/View/CompetitionMainView.swift
@@ -85,93 +85,3 @@ struct CompetitionMainView: View {
         }
     }
 }
-
-struct CreateRoomModalView: View {
-    @Environment(\.presentationMode) var presentationMode: Binding<PresentationMode>
-    @EnvironmentObject var competitionMainViewModel: CompetitionMainViewModel
-    @State var title: String = ""
-    @State var startDate: Date = Date()
-    @State var participants: Int = 2
-    @State var goal: String = "7"
-    
-    var body: some View {
-        Group {
-            TextField("Set room Title", text: $title)
-            HStack {
-                Text("Start Date")
-                DatePicker("",
-                           selection: $startDate,
-                           in: Date()..., displayedComponents: .date)
-            }
-            HStack {
-                Text("Goal")
-                TextField("", text: $goal)
-                    .textContentType(.telephoneNumber)
-                    .frame(width: 100)
-                    .background(Color.gray)
-            }
-            
-            HStack {
-                Text("Participants")
-                Spacer()
-                Button {
-                    participants = max(2, participants - 1)
-                } label: {
-                    Image(systemName: "minus.circle")
-                }
-                Text("\(participants)")
-                Button {
-                    participants = min(6, participants + 1)
-                } label: {
-                    Image(systemName: "plus.circle")
-                }
-            }
-            HStack {
-                Button {
-                    competitionMainViewModel.makeRoom(title: title,
-                                                      goal: Int(goal) ?? 0,
-                                                      maxParticipants: participants,
-                                                      startDate: startDate)
-                } label: {
-                    Text("Save")
-                }
-                Spacer()
-                Button {
-                    self.presentationMode.wrappedValue.dismiss()
-                } label: {
-                    Text("Dismiss")
-                }
-            }
-        }
-        .padding()
-    }
-}
-
-struct JoinRoomModalView: View {
-    @Environment(\.presentationMode) var presentationMode: Binding<PresentationMode>
-    @EnvironmentObject var competitionMainViewModel: CompetitionMainViewModel
-    @State var roomNumber: String = ""
-    
-    var body: some View {
-        VStack {
-            HStack {
-                Text("Room Number")
-                Spacer()
-                TextField("", text: $roomNumber)
-                    .textContentType(.telephoneNumber)
-                    .frame(width: 100)
-                    .background(Color.gray)
-            }
-            Button {
-                competitionMainViewModel.joinRoom(roomNumber)
-                if competitionMainViewModel.isJoinable {
-                    self.presentationMode.wrappedValue.dismiss()
-                }
-            } label: {
-                Text("Join")
-            }
-            Text(competitionMainViewModel.joinError)
-                .foregroundColor(.red)
-        }
-    }
-}
