+++ Git-Challenges/Let's Git it!/ViewModel/CompetitionMainViewModel.swift
@@ -39,7 +39,7 @@ class CompetitionMainViewModel: ObservableObject {
         }
     }
     
-    func makeRoom(title: String, goal: Int, maxParticipants: Int, startDate: Date) {
+    func createRoom(title: String, goal: Int, maxParticipants: Int, startDate: Date) {
         validRoomIDtoMake() { roomID in
             let db = Firestore.firestore()
             let userID = UserDefaults.shared.string(forKey: "userId") ?? Auth.auth().currentUser?.uid ?? "none"
@@ -50,9 +50,13 @@ class CompetitionMainViewModel: ObservableObject {
                                     participants: [userID],
                                     maxParticipants: maxParticipants).asDictionary!
             db.collection("RoomData").document(roomID).setData(roomData)
+            
+            // MARK: Reload Room Dats
+            
             self.getRoomDatas()
         }
     }
+    
     func joinRoom(_ roomNumber: String) {
         validRoomIDtoJoin(id: roomNumber) { isDone, errString in
             if isDone == false {
@@ -63,69 +67,5 @@ class CompetitionMainViewModel: ObservableObject {
                 self.isJoinable = true
             }
         }
-    }
-
-    func validRoomIDtoJoin(id: String, completionHandler: @escaping (Bool, String)-> Void) {
-        
-        // MARK: Check Room ID  with Completion for async
-        
-        let db = Firestore.firestore()
-        let docRef = db.collection("RoomData").document(id)
-        let userID = UserDefaults.shared.string(forKey: "userId") ?? Auth.auth().currentUser?.uid ?? "none"
-        
-        docRef.getDocument { document, err in
-            if let document = document, document.exists == true {
-                let maxParticipants = document.get("maxParticipants") as! Int
-                let currentParticipants = document.get("participants") as! [String]
-                let numberParticipants = currentParticipants.count
-                let isAlreadyJoined = currentParticipants.contains(userID)
-                
-                if numberParticipants >= maxParticipants {
-                    completionHandler(false, "Room is full.")
-                } else if isAlreadyJoined {
-                    completionHandler(false, "Already in The Room.")
-                } else {
-                    docRef.updateData(["participants": FieldValue.arrayUnion([userID])])
-                    completionHandler(true, " ")
-                }
-            } else {
-                completionHandler(false, "Room doesn't exist.")
-            }
-        }
-    }
-}
-
-func validRoomIDtoMake(completionHandler: @escaping (String)-> Void) {
-    let id: String = randomSixDigitCode()
-    
-    // MARK: Random Room ID Generate with Completion for async
-    
-    let db = Firestore.firestore()
-    let docRef = db.collection("RoomData").document(id)
-    docRef.getDocument { document, err in
-        if let document = document, document.exists == true {
-            validRoomIDtoMake(completionHandler: completionHandler)
-        } else {
-            completionHandler(id)
-        }
-    }
-}
-
-func randomSixDigitCode() -> String {
-    var number = String()
-    for _ in 1...6 {
-        number += "\(Int.random(in: 1...9))"
-    }
-    return number
-}
-
-extension Encodable {
-    /// Object to Dictionary
-    /// cf) Dictionary to Object: JSONDecoder().decode(Object.self, from: dictionary)
-    var asDictionary: [String: Any]? {
-        guard let object = try? JSONEncoder().encode(self),
-              let dictinoary = try? JSONSerialization.jsonObject(with: object, options: [])
-                as? [String: Any] else { return nil }
-        return dictinoary
     }
 }
